pyproject.toml
[project]
name = "python-agent"
version = "0.1.0"
description = "Testing GEMINI API"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
